Basic support for 118 countries, and 780. Europe and North America are tested

+++ src/CareLink/SupportClasses/DrawingExtensions.vb
@@ -6,7 +6,7 @@ Imports System.Windows.Forms.DataVisualization.Charting
 
 Module DrawingExtensions
 
-    Private Function GetColorFromTimeToNextCalib(timeToNextCalibHours As Byte) As Color
+    Private Function GetColorFromTimeToNextCalib(timeToNextCalibHours As UShort) As Color
         If timeToNextCalibHours <= 2 Then
             Return Color.Red
         ElseIf timeToNextCalibHours < 4 Then
@@ -16,7 +16,7 @@ Module DrawingExtensions
         End If
     End Function
 
-    Friend Function DrawCenteredArc(backImage As Bitmap, TimeToNextCalibration As Byte, arcPercentage As Double, Optional colorTable As IReadOnlyDictionary(Of String, Color) = Nothing, Optional segmentName As String = "") As Bitmap
+    Friend Function DrawCenteredArc(backImage As Bitmap, TimeToNextCalibration As UShort, arcPercentage As Double, Optional colorTable As IReadOnlyDictionary(Of String, Color) = Nothing, Optional segmentName As String = "") As Bitmap
         If arcPercentage < Double.Epsilon Then
             Return backImage
         End If
